weekly update 25/01/2015
Changed the format of the sqlite DB.
Added methods to get bets for charts.
Updated the Graph Form to be a bit more efficient.
Added the ability to upload the chart to imgur on save
added live charts.
Fixed simulations
Fixed bug of resetting after withdrawal of investment.
Added 0 betting/ prebetting
added error messgaes for JD and PD