bet history form, new embedded live chart, other minor change
Cleaned up a bit of code. Added a form where you can view previous bets.
Changed the high column in the db to a smallint instead of bit because
sqlite returns false for bit regardless of the value. replaced some text
boxes with numeric updowns.  added a embedded live chart to the main
form.